+++ create_package.sh
@@ -13,23 +13,17 @@ COMPOSER_VERSION=$(cat composer.json \
   | sed 's/[",]//g' \
   | tr -d '[[:space:]]')
 
-PACKAGE_VERSION=$(xmllint etc/module.xml --xpath 'string(//config/module/@setup_version)')
+echo "Creating package..."
+echo "Version: ${bold}$PACKAGE_VERSION ${normal}"
 
-if [ $COMPOSER_VERSION == $PACKAGE_VERSION ]; then
-    echo "Creating package..."
-    echo "Version: ${bold}$PACKAGE_VERSION ${normal}"
+rm -rf bm-bluepayment-*.zip
 
-    rm -rf bm-bluepayment-*.zip
+zip -r "bm-bluepayment-$PACKAGE_VERSION.zip" ./ \
+  -x *.idea* \
+  -x *.git* \
+  -x *.DS_Store* \
+  -x *create_package.sh* \
+  -x *.doc*
 
-    zip -r "bm-bluepayment-$PACKAGE_VERSION.zip" ./ \
-      -x *.idea* \
-      -x *.git* \
-      -x *.DS_Store* \
-      -x *create_package.sh* \
-      -x *.doc*
-
-    echo "======================================================================================================"
-    echo "${green}Package ${bold}bm-bluepayment-$PACKAGE_VERSION.zip${normal}${green} created"
-else
-    echo "${red}Composer version ${bold}$COMPOSER_VERSION${normal}${red} is not equal to package version ${bold}$PACKAGE_VERSION${normal}"
-fi
+echo "======================================================================================================"
+echo "${green}Package ${bold}bm-bluepayment-$PACKAGE_VERSION.zip${normal}${green} created"